+++ Y4C/Y4C/Controllers/HomeController.cs
@@ -40,32 +40,10 @@ namespace Y4C.Controllers
             }
             return RedirectToAction(nameof(ThankYou), new { id = add.Id });
         }
-        public ActionResult ManageContent()
-        {
-            return View(DBcontext.AC.ToList());
-        }
+        
         public ActionResult ThankYou()
         {
             return View();
-        }
-
-        public ActionResult ContentDetails(int id=0)
-        {
-            return View(DBcontext.AC.Find(id));
-        }
-
-        public ActionResult DeleteContent(int id=0)
-        {
-            return View(DBcontext.AC.Find(id));
-        }
-
-        [HttpPost, ActionName("DeleteContent")]
-        public ActionResult DeleteConfirm(int id)
-        {
-            AddContent post = DBcontext.AC.Find(id);
-            DBcontext.AC.Remove(post);
-            DBcontext.SaveChanges();
-            return RedirectToAction("ManageContent");
         }
         //public ViewResult PlayVideo() => View();
