+++ Y4C/Y4C/Views/Home/AddContent.cshtml
@@ -33,6 +33,8 @@
                     <input type="text" class="form-control" name="Audio" placeholder="Enter the url of the themes audio" />
                 </div>
 
+                <input type="hidden" name="ArchiveStatus" value="Active" />
+
                 <h2>Upload Blog</h2>
 
                 <textarea id="summernote" name="BlogContent"></textarea>
