+++ Y4C/Y4C/Views/Home/HomePage.cshtml
@@ -9,7 +9,8 @@
 <div class="container">
     @foreach (var item in @Model)
     {
-
+        if (item.ArchiveStatus == "Active")
+        {
         <div class="col-sm-6 col-md-4 col-lg-3 mt-4">
             <div class="card">
                 <div class="card" style="width: 30rem">
@@ -22,6 +23,7 @@
                 </div>
             </div>
         </div>
+        }
     }
 </div>
 <br />
