--- Y4C/Y4C/Views/Home/ManageContent.cshtml
@@ -1,39 +1,0 @@
-﻿@model System.Collections.Generic.IEnumerable<Y4C.Models.AddContent>
-@{
-    ViewData["Title"] = "ManageContent";
-}
-
-<h2>ManageContent</h2>
-
-
-<table class="table">
-    <thead>
-        <tr>
-            <th>
-                @Html.DisplayNameFor(model => model.Title)
-            </th>
-            <th>
-                @Html.DisplayNameFor(model => model.Description)
-            </th> 
-            <th></th>
-        </tr>
-    </thead>
-    <tbody>
-        @foreach (var item in @Model)
-        {
-            <tr>
-                <td>
-                    @Html.DisplayFor(modelItem => item.Title)
-                </td>
-                <td>
-                    @Html.DisplayFor(modelItem => item.Description)
-                </td>
-                <td>
-                    <a asp-action="Edit" asp-route-id="@item.Id">Edit</a> |
-                    <a asp-action="ContentDetails" asp-route-id="@item.Id">Details</a> |
-                    <a asp-action="DeleteContent" asp-route-id="@item.Id">Delete</a>
-                </td>
-            </tr>
-        }
-    </tbody>
-</table>
