+++ Y4C/Y4C/Views/Manage/Archive.cshtml
@@ -1,0 +1,79 @@
+﻿@model Y4C.Models.AddContent
+
+@{
+    ViewData["Title"] = "Archive";
+}
+
+<h2>Archive</h2>
+
+<h3>Are you sure you want to archive this post?</h3>
+<hr />
+<div class="row">
+    <div class="col-md-4">
+        <form asp-action="Archive">
+            <div asp-validation-summary="ModelOnly" class="text-danger"></div>
+            <input type="hidden" asp-for="Id" />
+            <input type="hidden" asp-for="Title" />
+            <input type="hidden" asp-for="Video" />
+            <input type="hidden" asp-for="Description" />
+            <input type="hidden" asp-for="Audio" />
+            <input type="hidden" asp-for="BlogContent" />
+            <input type="hidden" asp-for="BlogAuthor" />
+            <!--
+    <div class="form-group">
+        <label asp-for="Title" class="control-label"></label>
+        <input asp-for="Title" class="form-control" />
+        <span asp-validation-for="Title" class="text-danger"></span>
+    </div>
+    <div class="form-group">
+        <label asp-for="Video" class="control-label"></label>
+        <input asp-for="Video" class="form-control" />
+        <span asp-validation-for="Video" class="text-danger"></span>
+    </div>
+    <div class="form-group">
+        <label asp-for="Description" class="control-label"></label>
+        <input asp-for="Description" class="form-control" />
+        <span asp-validation-for="Description" class="text-danger"></span>
+    </div>
+    <div class="form-group">
+        <label asp-for="Audio" class="control-label"></label>
+        <input asp-for="Audio" class="form-control" />
+        <span asp-validation-for="Audio" class="text-danger"></span>
+    </div>
+    <div class="form-group">
+        <label asp-for="BlogContent" class="control-label"></label>
+        <input asp-for="BlogContent" class="form-control" />
+        <span asp-validation-for="BlogContent" class="text-danger"></span>
+    </div>
+    <div class="form-group">
+        <label asp-for="BlogTitle" class="control-label"></label>
+        <input asp-for="BlogTitle" class="form-control" />
+        <span asp-validation-for="BlogTitle" class="text-danger"></span>
+    </div>
+    <div class="form-group">
+        <label asp-for="BlogAuthor" class="control-label"></label>
+        <input asp-for="BlogAuthor" class="form-control" />
+        <span asp-validation-for="BlogAuthor" class="text-danger"></span>
+    </div>
+        -->
+
+            <div class="form-group">
+                <label class="control-label">Change Archive Status:</label>
+                <select class="form-control" name="ArchiveStatus">
+                    <option disabled selected="selected">Choose Status</option>
+                    <option value="Active" >Active</option>
+                    <option value="Inactive" >Inactive</option>
+                </select>
+                <span asp-validation-for="ArchiveStatus" class="text-danger"></span>
+            </div>
+            <div class="form-group">
+                <input type="submit" value="Save" class="btn btn-default" />
+            </div>
+        </form>
+    </div>
+</div>
+
+<div>
+    <a asp-action="ManageContent">Cancel</a>
+</div>
+
