+++ Y4C/Y4C/Views/Manage/DeleteContent.cshtml
