+++ Y4C/Y4C/Views/Shared/_Layout.cshtml
@@ -44,7 +44,7 @@
                         </a>
                         <ul class="dropdown-menu">
                             <li><a asp-area="" asp-controller="Home" asp-action="AddContent">Upload</a></li>
-                            <li><a asp-area="" asp-controller="Home" asp-action="ManageContent">Manage</a></li>
+                            <li><a asp-area="" asp-controller="Manage" asp-action="ManageContent">Manage</a></li>
                         </ul>
                     </li>
                 </ul>
